+++ samples/International.EInvoicing.Samples/Program.cs
@@ -20,6 +20,7 @@ Extending.Run();
 HybridPdf.Run(Invoices.Build(announce: false));
 
 CountryInvoices.Run(einvoicing);
+CountryShortcuts.Run();
 FrenchLifecycle.Run(einvoicing);
 FrenchEReporting.Run();
 NationalIdentifiers.Run();